글자모양-음영색에 색상없음 설정 안되는 오류수정..

+++ src/main/java/kr/dogfoot/hwp2hwpx/section/object/gso/ForPicture.java
@@ -107,8 +107,8 @@ public class ForPicture extends ForShapeComponent {
     private void imgDim() {
         picture.createImgDim();
         picture.imgDim()
-                .dimwidthAnd((int) hwpSCP.getImageWidth())
-                .dimheight((int) hwpSCP.getImageHeight());
+                .dimwidthAnd(hwpSCP.getImageWidth())
+                .dimheight(hwpSCP.getImageHeight());
     }
 
     private void img() {
